+++ lib/kitchen/exercise_element.rb
@@ -38,5 +38,13 @@ module Kitchen
     def solution
       first("div[data-type='solution']")
     end
+
+    # Returns whether the exercise has been baked
+    #
+    # @return [Boolean]
+    #
+    def baked?
+      search('div.os-problem-container').any?
+    end
   end
 end